+++ packages/nesting/stylelint-mrhenry-nesting.test.js
@@ -21,7 +21,7 @@ testRule({
 		},
 		{
 			code: "@custom-selector :--foo bar;",
-			description: "Regular conditional media rule"
+			description: "Regular at rule"
 		},
 		{
 			code: "div { @supports (display: grid) { color: green; } }",
@@ -38,6 +38,26 @@ testRule({
 		{
 			code: "div { &:not(.foo), &:is(.bar) { color: green; } }",
 			description: "Functional pseudo class selector"
+		},
+		{
+			code: "div { &:is(body.theme-red *) { color: red } }",
+			description: "Alternative to multiple &, in pseudo functions",
+		},
+		{
+			code: "div { &:is(body.theme-red &) { color: red } }",
+			description: "Multiple &, in pseudo functions",
+		},
+		{
+			code: "div { &:is(&&) { color: red } }",
+			description: "Multiple &, in pseudo functions",
+		},
+		{
+			code: "div { &:is(& + &) { color: red } }",
+			description: "Multiple &, in pseudo functions",
+		},
+		{
+			code: "div { { color: red; } }",
+			description: "empty selector",
 		},
 	],
 
@@ -97,13 +117,31 @@ testRule({
 			endColumn: 18
 		},
 		{
-			code: "div { &:not(&) { color: red } }",
-			description: "Multiple &, in pseudo functions",
-			message: rule.messages.rejectedMustContainOnlyOneAmpersand(),
+			code: "div { && { color: red } }",
+			description: "Multiple & top level",
+			message: rule.messages.rejectedMustEndWithPseudo(),
 			line: 1,
 			column: 7,
 			endLine: 1,
-			endColumn: 15
+			endColumn: 9
+		},
+		{
+			code: "div { & + & { color: red } }",
+			description: "Multiple & top level",
+			message: rule.messages.rejectedNestingSelectorIncorrectShape(),
+			line: 1,
+			column: 7,
+			endLine: 1,
+			endColumn: 12
+		},
+		{
+			code: "div { & &:focus { color: red } }",
+			description: "Multiple & top level",
+			message: rule.messages.rejectedNestingSelectorIncorrectShape(),
+			line: 1,
+			column: 7,
+			endLine: 1,
+			endColumn: 16
 		},
 		{
 			code: "div { & > bar + .bar { color: red } }",
@@ -202,5 +240,133 @@ testRule({
 			endLine: 1,
 			endColumn: 14
 		},
+		{
+			code: ".bar { body.theme-red & { color: magenta; } }",
+			fixed: ".bar { &:is(body.theme-red *) { color: magenta; } }",
+			description: "Incorrect shape",
+			message: rule.messages.rejectedMustStartWithAmpersand(),
+			line: 1,
+			column: 8,
+			endLine: 1,
+			endColumn: 24
+		},
+		{
+			code: ".bar { body.theme-red > & { color: magenta; } }",
+			fixed: ".bar { &:is(body.theme-red > *) { color: magenta; } }",
+			description: "Incorrect shape",
+			message: rule.messages.rejectedMustStartWithAmpersand(),
+			line: 1,
+			column: 8,
+			endLine: 1,
+			endColumn: 26
+		},
+		{
+			code: ".bar { body.theme-red + & { color: magenta; } }",
+			fixed: ".bar { &:is(body.theme-red + *) { color: magenta; } }",
+			description: "Incorrect shape",
+			message: rule.messages.rejectedMustStartWithAmpersand(),
+			line: 1,
+			column: 8,
+			endLine: 1,
+			endColumn: 26
+		},
+		{
+			code: ".bar { body.theme-red ~ & { color: magenta; } }",
+			fixed: ".bar { &:is(body.theme-red ~ *) { color: magenta; } }",
+			description: "Incorrect shape",
+			message: rule.messages.rejectedMustStartWithAmpersand(),
+			line: 1,
+			column: 8,
+			endLine: 1,
+			endColumn: 26
+		},
+		{
+			code: ".bar { body.theme-red& { color: magenta; } }",
+			fixed: ".bar { &:is(body.theme-red&) { color: magenta; } }",
+			description: "Incorrect shape",
+			message: rule.messages.rejectedMustStartWithAmpersand(),
+			line: 1,
+			column: 8,
+			endLine: 1,
+			endColumn: 23
+		},
+		{
+			code: ".bar { :focus& { color: magenta; } }",
+			fixed: ".bar { &:focus { color: magenta; } }",
+			description: "Incorrect shape",
+			message: rule.messages.rejectedMustStartWithAmpersand(),
+			line: 1,
+			column: 8,
+			endLine: 1,
+			endColumn: 15
+		},
+		{
+			code: ".bar { body.theme-red .bar { color: magenta; } }",
+			fixed: ".bar { &:is(body.theme-red .bar) { color: magenta; } }",
+			description: "Incorrect shape",
+			message: rule.messages.rejectedMustStartWithAmpersand(),
+			line: 1,
+			column: 8,
+			endLine: 1,
+			endColumn: 27
+		},
+		{
+			code: ".bar { body.theme-red * { color: magenta; } }",
+			fixed: ".bar { &:is(body.theme-red *) { color: magenta; } }",
+			description: "Incorrect shape",
+			message: rule.messages.rejectedMustStartWithAmpersand(),
+			line: 1,
+			column: 8,
+			endLine: 1,
+			endColumn: 24
+		},
 	]
+});
+
+testRule({
+	plugins: ["./stylelint-mrhenry-nesting.js"],
+	ruleName,
+	config: [true, { ignoreAtRules: ['unknown', /^YOUR-/i] }],
+
+	accept: [
+		{
+			code: "div { @unknown foo; }",
+			description: "Ignore unknown at rule by string"
+		},
+		{
+			code: "div { @yOuR-rule foo; }",
+			description: "Ignore unknown at rule by regexp"
+		},
+	],
+
+	reject: [
+		{
+			code: "div { @not-yOuR-rule foo; }",
+			description: "@custom-selector",
+			message: rule.messages.rejectedAtRule('not-your-rule'),
+			line: 1,
+			column: 7,
+			endLine: 1,
+			endColumn: 20
+		},
+	]
+});
+
+testRule({
+	plugins: ["./stylelint-mrhenry-nesting.js"],
+	ruleName,
+	config: [false],
+
+	accept: [
+		{
+			code: "div { @unknown foo; }",
+			description: "Ignore unknown at rule by string"
+		},
+		{
+			code: "div { @yOuR-rule foo; }",
+			description: "Ignore unknown at rule by regexp"
+		},
+	],
+
+	reject: []
 });