+++ packages/prop-order/stylelint-mrhenry-prop-order.js
@@ -20,6 +20,10 @@ const ignoredAtRules = [
 
 const ruleFunction = (primaryOption, secondaryOptionObject, context) => {
 	return (postcssRoot, postcssResult) => {
+		if (!primaryOption) {
+			return;
+		}
+		
 		postcssRoot.walkRules((rule) => {
 			let parent = rule.parent;
 			while (parent) {