+++ packages/nesting/CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
+## 2.1.0
+
+- Allow multiple `&` as long as the general pattern remains (e.g. `&:is(.bar &)`)
+- Add `ignoreAtRules` plugin option so that non-standard things like `@mixins` can be allowed.
+- Improve auto fixing.
+- Fix plugin option `false`
+
 ## 2.0.0
 
-Change rule name to `@mrhenry/stylelint-mrhenry-nesting`
+- Change rule name to `@mrhenry/stylelint-mrhenry-nesting`
 
 ## 1.0.3